refactor(cockpit): extract _AlertDetailModal into sibling module (#1354)
Fourth wedge of the cockpit_ui.py god-module split (after #1606, #1618,
#1670). The rail's alert-detail modal (the read + recover surface for
the ``♡⚠`` badge added in #989) is the next-cleanest self-contained
class: 183 LOC of pure Textual widget with no cockpit state or service
dependencies, used only from cockpit_ui.py and not referenced by any
test or other module.

Moves the class to a new sibling module src/pollypm/cockpit_alert_detail.py
with the same docstring + contract block the prior wedges adopted.
cockpit_ui.py keeps an import shim so the existing
``from pollypm.cockpit_ui import _AlertDetailModal`` surface is
preserved (the rail/action wiring at lines 1025/1228 stays unchanged).

cockpit_ui.py: 18937 -> 18757 LOC (-180); diff in cockpit_ui.py is 186
lines (183 deletions + 3-line import shim).
new cockpit_alert_detail.py: 220 LOC (incl. module + class docstrings).
